Rename and reorganise some of the map related data files
